requirements.txt
# Core data processing
pandas==2.3.2
numpy==2.3.3
scipy==1.16.2
python-dotenv==1.1.1
beautifulsoup4==4.13.5
requests==2.32.5
scrapy==2.11.0
lxml==6.0.2
fredapi==0.5.2
yfinance==0.2.66
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
multitasking==0.0.12
peewee==3.18.2
curl-cffi==0.13.0
platformdirs==4.4.0
websockets==15.0.1
protobuf==6.32.1
frozendict==2.4.6
xyzservices==2025.4.0
branca==0.8.2

# Machine learning & optimisation
scikit-learn==1.7.2
deap==1.4.3
xgboost==3.0.5
shap==0.48.0
networkx==3.3
python-louvain==0.16  # For Louvain community detection (Akinyede & Caruso 2025)
numba==0.62.1
llvmlite==0.45.1
tqdm==4.67.1
cloudpickle==3.1.1
slicer==0.0.8
matplotlib==3.9.2
seaborn==0.13.2
textblob==0.18.0.post0
nltk==3.9.1
plotly==6.3.0
kaleido==0.2.1
dash==2.14.1
folium==0.20.0
shapely==2.1.2
Flask-SocketIO==5.5.1
redis==6.4.0
celery==5.4.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.9
flask==2.3.3
flask-cors==6.0.1
werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
cachelib==0.13.0
Flask-Caching==2.3.1

# Testing & tooling
pytest==8.3.3
black==24.8.0
flake8==7.1.1
